Rewrite default interceptors to include proper header
